requirements.txt
Django
Pillow
numpy
scikit-image
tensorflow
huggingface-hub